Kaggle Spaceship Titanic Competition
This was part of a group assignment for my Business Challenge at Hult International Business School. My group, where I acted as a leader, did a classification analysis by creating 9 different models that predicted our y-variable, "Transported". Before anything, we cleaned the dataset by dealing with missing values, transforming categorical values into 1 and 0, as well as feature engineering and hyperparameter tunning the models in order to get higher accuracy scores, which put us in the top 22% of a total of 2,306 teams worldwide on Kaggle.